+++ app/src/androidTest/java/com/github/orkest/ViewModel/auth/MockAuthViewModel.kt
@@ -7,7 +7,7 @@ class MockAuthViewModel : AuthViewModel() {
     companion object {
         val EXISTING_USER = "exists"
         val VALID_USER = "valid"
-        val NO_PERMISSIONS = "no_permissions"
+        val INVALID_PAIR = "invalid pair"
         const val EMPTY_USER = ""
         const val NO_CONNECTION = "No wifi"
     }
@@ -38,13 +38,21 @@ class MockAuthViewModel : AuthViewModel() {
     @Override
     override fun signInUser(): CompletableFuture<Boolean> {
         val future = CompletableFuture<Boolean>()
-        val username = getCurrentUsername().text
+        val username = getUsername().text
 
         if (username == EXISTING_USER)
             future.complete(true)
 
-        if (username == NO_PERMISSIONS)
+        if (username == INVALID_PAIR)
             future.complete((false))
+
+        if (username == EMPTY_USER)
+            future.completeExceptionally(Exception("Username cannot be empty!"))
+
+        if (username == NO_CONNECTION)
+            future.completeExceptionally(
+                Exception("Sorry, something went wrong ... " +
+                        "Please check your Internet connection"))
 
         return future
     }
